+++ test/unittests/unit_list.c
@@ -40,7 +40,7 @@ int __wrap_printf(const char *format, ...) {
     return 1;
 }
 
-list_t testlist;
+static list_t testlist = {.element_prealloc_count = 0};
 
 static void test_contains(void **state) {